test: consolidate all fakes under tests/fakes/ (one file per fake)
Move every inline ``Fake*`` class out of ``tests/conftest.py`` into its
own file under ``tests/fakes/``. ``tests/conftest.py`` retains only
fixture wiring plus the ``_make_retry`` / ``_make_testable_plugin``
helpers; scales the convention started in #685 across the whole suite.

Fakes moved (each gets its own ``tests/fakes/fake_<snake>.py``):

- FakeStatePersister
- FakeSettingsPersister
- FakeMetadataCachePersister
- FakeSaveSyncStatePersister
- FakePluginMetadataReader
- FakeFirmwareCachePersister
- FakeCoverArtFileStore
- FakeSgdbArtworkCache
- FakeDownloadFileStore
- FakeMigrationFileStore
- FakeDownloadQueueCleanup
- FakeDownloadQueueAdapter
- FakeFirmwareFileStore
- FakeRomFileStore
- FakeSaveFileStore
- FakePathProbe
- FakeHostnameProvider
- FakeCoreInfoProvider
- FakeRetroDeckPaths

Import-site updates: every ``from conftest import Fake*`` switches to
``from fakes.fake_<snake> import Fake*``. The ``importlib.import_module``
re-export dance in ``tests/services/library/conftest.py`` collapses
into a direct ``from fakes.*`` import. Behavior is preserved — full
suite (2508 tests) stays green.

``tests/conftest.py``: 1005 LOC → 122 LOC.

Closes #693
