+++ tests/services/test_rom_removal.py
@@ -7,7 +7,9 @@ import sys
 from unittest.mock import MagicMock
 
 import pytest
-from conftest import FakeDownloadQueueCleanup, FakeRetroDeckPaths, FakeRomFileStore
+from fakes.fake_download_queue_cleanup import FakeDownloadQueueCleanup
+from fakes.fake_retrodeck_paths import FakeRetroDeckPaths
+from fakes.fake_rom_file_store import FakeRomFileStore
 
 sys.path.insert(0, os.path.join(os.path.dirname(__file__), "py_modules"))
 sys.path.insert(0, os.path.dirname(__file__))
