+++ tests/test_plugin.py
@@ -4,14 +4,12 @@ import os
 from unittest.mock import MagicMock
 
 import pytest
-from conftest import (
-    FakeMetadataCachePersister,
-    FakePathProbe,
-    FakeRetroDeckPaths,
-    FakeSettingsPersister,
-    FakeSgdbArtworkCache,
-    FakeStatePersister,
-)
+from fakes.fake_metadata_cache_persister import FakeMetadataCachePersister
+from fakes.fake_path_probe import FakePathProbe
+from fakes.fake_retrodeck_paths import FakeRetroDeckPaths
+from fakes.fake_settings_persister import FakeSettingsPersister
+from fakes.fake_sgdb_artwork_cache import FakeSgdbArtworkCache
+from fakes.fake_state_persister import FakeStatePersister
 from fakes.system_time import FakeClock, FakeSleeper, FakeUuidGen
 
 from adapters.debug_logger import SettingsAwareDebugLogger
